Move similar functions to helpers
