+++ src/layouts/InterfaceLayout/containers/InteractWithContractContainer/InteractWithContractContainer.vue
@@ -275,7 +275,6 @@ import InterfaceContainerTitle from '../../components/InterfaceContainerTitle';
 import InterfaceBottomText from '@/components/InterfaceBottomText';
 import { Misc, Toast } from '@/helpers';
 import { isAddress } from '@/helpers/addressUtils';
-import { uint, address, string, bytes, bool } from '@/helpers/solidityTypes.js';
 import * as unit from 'ethjs-unit';
 import store from 'store';
 
@@ -372,42 +371,7 @@ export default {
       this.value = 0;
       this.inputs = {};
     },
-    isValidInput(value, solidityType) {
-      if (!value) value = '';
-      if (solidityType.includes('[') && solidityType.includes(']')) {
-        const values = [];
-        if (value[0] === '[') {
-          const strToArr =
-            value[0] === '[' ? value.substr(0, value.length - 1) : value;
-          strToArr
-            .replace(/\s/, '')
-            .split(',')
-            .forEach(item => {
-              if (solidityType.includes(uint)) {
-                values.push(value !== '' && !isNaN(value) && Misc.isInt(value));
-              } else if (solidityType.includes(address)) {
-                values.push(isAddress(value));
-              } else if (solidityType.includes(string)) {
-                values.push(isAddress(true));
-              } else if (solidityType.includes(bool)) {
-                values.push(typeof value === typeof true || value === '');
-              } else if (solidityType.includes(bytes)) {
-                values.push(Misc.validateHexString(item));
-              }
-            });
-        }
-        return !values.includes(false);
-      }
-      if (solidityType === 'uint')
-        return value !== '' && !isNaN(value) && Misc.isInt(value);
-      if (solidityType === 'address') return isAddress(value);
-      if (solidityType === 'string') return true;
-      if (solidityType === 'bytes')
-        return value.substr(0, 2) === '0x' && Misc.validateHexString(value);
-      if (solidityType === 'bool')
-        return typeof value === typeof true || value === '';
-      return false;
-    },
+    isValidInput: Misc.isContractArgValid,
     getType: Misc.solidityType,
     selectedContract(selected) {
       if (selected.abi === '') {
